+++ game/chess.py
@@ -1,4 +1,5 @@
 from game.board import Board
+from game.exceptions import EmptyPosition, InvalidMove, InvalidTurn
 
 class Chess:
     def __init__(self):
@@ -16,8 +17,14 @@ class Chess:
         to_col,
     ):
         # validate coords
-        self.__board__.get_piece(from_row, from_col)
-        self.__board__.move_piece(from_row, from_col, to_row, to_col)
+        piece = self.__board__.get_piece(from_row, from_col)
+        if not piece:
+            raise EmptyPosition()
+        if not piece.get_color() == self.__turn__:
+            raise InvalidTurn()
+        if not piece.valid_positions(from_row, from_col, to_row, to_col):
+            raise InvalidMove()
+        self.__board__.move(from_row, from_col, to_row, to_col)
         self.change_turn()
 
     @property