+++ tests/test_rook.py
@@ -21,19 +21,19 @@ class TestRook(unittest.TestCase):
         self.assertEqual(possibles, [(5,1),(6,1),(7,1)])
 
     def test_move_vertical_asc(self):
-        board = Board()
+        board = Board(for_test=True)
         rook = Rook('WHITE',board)
-        possibles = rook.possible_positions_va(4,1)
-        self.assertEqual(possibles, [(3,1),(2,1),(1,1),(0,1)])
+        possibles = rook.possible_positions_va(4,2)
+        self.assertEqual(possibles, [(3,2),(2,2),(1,2),(0,2)])
 
     def test_move_horizontal_right(self):
-        board = Board()
+        board = Board(for_test=True)
         rook = Rook('WHITE',board)
         possibles = rook.possible_positions_hr(4,4)
         self.assertEqual(possibles, [(4,5),(4,6),(4,7)])
 
     def test_move_horizontal_left(self):
-        board = Board()
+        board = Board(for_test=True)
         rook = Rook('WHITE',board)
         possibles = rook.possible_positions_hl(4,4)
         self.assertEqual(possibles, [(4,3),(4,2),(4,1),(4,0)])