+++ changelog.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.10] - 2024-09-05
+## [0.1.11] - 2024-09-08
+### Added
+
+- exceptions.py, where all exceptions are defined
+
+- applied exceptions to chess
+
+## [0.1.10] - 2024-09-07
 ### Added
 
 - ortogonal verifications for color of pieces in rook